+++ lib/styles/card_item_style.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 class SelectPickerCardItemStyle {
   SelectPickerCardItemStyle({
-    this.dividerColor = Colors.black87,
+    this.dividerColor = Colors.grey,
     this.backgroundColor = Colors.white,
     this.textColor = Colors.black87,
     this.dividerHeight = 1,