+++ lib/widgets/body/select_picker_body.dart
@@ -1,11 +1,10 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:provider/provider.dart';
 import 'package:selectpicker/styles/bottomsheet_style.dart';
 import 'package:selectpicker/styles/card_item_style.dart';
 import 'package:selectpicker/styles/inputsearch_style.dart';
 import 'package:selectpicker/models/select_picker_item.dart';
+import 'package:selectpicker/styles/top_selector_style.dart';
 import 'package:selectpicker/widgets/body/card_select_picker.dart';
 import 'package:selectpicker/widgets/input/input_no_border.dart';
 import 'package:selectpicker/selectpicker_viewmodel.dart';
@@ -20,6 +19,7 @@ class SelectPickerBody extends StatefulWidget {
     required this.selectPickerBottomSheetStyle,
     required this.selectPickerInputSearchStyle,
     required this.selectPickerCardItemStyle,
+    required this.topSelectorStyle,
     this.showId = false,
   });
 
@@ -31,6 +31,7 @@ class SelectPickerBody extends StatefulWidget {
   final SelectPickerBottomSheetStyle selectPickerBottomSheetStyle;
   final SelectPickerInputSearchStyle selectPickerInputSearchStyle;
   final SelectPickerCardItemStyle selectPickerCardItemStyle;
+  final SelectPickerTopSelectorStyle topSelectorStyle;
 
   @override
   State<SelectPickerBody> createState() => _SelectPickerBodyState();
@@ -46,10 +47,11 @@ class _SelectPickerBodyState extends State<SelectPickerBody> {
         children: [
           header(),
           hintSearch(),
-          Container(
-            color: widget.selectPickerCardItemStyle.dividerColor ?? Colors.grey,
-            height: widget.selectPickerCardItemStyle.dividerHeight ?? 1,
-          ),
+          if (widget.selectPickerBottomSheetStyle.showSeparatorAboveTitle == true)
+            Container(
+              color: widget.selectPickerCardItemStyle.dividerColor ?? Colors.grey,
+              height: widget.selectPickerCardItemStyle.dividerHeight ?? 1,
+            ),
           Expanded(
             child: ListView.builder(
               shrinkWrap: false,
@@ -71,74 +73,126 @@ class _SelectPickerBodyState extends State<SelectPickerBody> {
   }
 
   Widget header() {
-    return Padding(
-      padding: const EdgeInsets.all(10),
+    return Ink(
+      height: widget.topSelectorStyle.height,
+      decoration: BoxDecoration(
+          color: widget.topSelectorStyle.backgroundColor ?? Theme.of(context).colorScheme.primary,
+          borderRadius: BorderRadius.only(
+            topRight: Radius.circular(widget.topSelectorStyle.borderRadius ?? 0),
+            topLeft: Radius.circular(widget.topSelectorStyle.borderRadius ?? 0),
+            bottomRight: Radius.circular((widget.topSelectorStyle.borderRadius ?? 1) / 4),
+            bottomLeft: Radius.circular((widget.topSelectorStyle.borderRadius ?? 1) / 4),
+          ),
+          border: Border.all(width: 3, color: widget.selectPickerBottomSheetStyle.backgroundColor ?? Colors.white)),
       child: Row(
+        crossAxisAlignment: CrossAxisAlignment.center,
         children: [
           Expanded(
             flex: 1,
-            child: Align(
-              alignment: FractionalOffset.topCenter,
-              child: Padding(
-                padding: const EdgeInsets.fromLTRB(0, 5, 20, 0),
-                child: Container(),
-              ),
-            ),
+            child: widget.topSelectorStyle.iconTitle ?? SizedBox(),
           ),
           Expanded(
-            flex: 10,
-            child: Align(
-              alignment: Alignment.center,
-              child: Padding(
-                padding: const EdgeInsets.only(top: 5),
-                child: Text(
-                  widget.hint,
-                  textAlign: TextAlign.left,
-                  style: TextStyle(
-                    color: widget.selectPickerInputSearchStyle.textColor,
-                    fontSize: 18,
-                    height: 1,
-                  ),
-                ),
-              ),
-            ),
-          ),
-          Expanded(
-            flex: 2,
+            flex: 6,
             child: Padding(
-              padding: const EdgeInsets.all(8.0),
-              child: Align(
-                alignment: Alignment.centerRight,
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    InkWell(
-                      borderRadius: BorderRadius.circular(50),
-                      onTap: () {
-                        Navigator.pop(context);
-                      },
-                      child: Center(
-                        child: Icon(
-                          Icons.close,
-                          color: widget.selectPickerBottomSheetStyle.closeIconColor,
-                          size: widget.selectPickerBottomSheetStyle.iconSize,
-                        ),
-                      ),
-                    ),
-                  ],
+              padding: EdgeInsets.only(top: (widget.topSelectorStyle.height ?? 1) / 10),
+              child: Text(
+                widget.hint,
+                textAlign: widget.topSelectorStyle.iconTitle == null ? TextAlign.center : TextAlign.left,
+                style: widget.topSelectorStyle.textStyle,
+              ),
+            ),
+          ),
+          Expanded(
+            flex: 1,
+            child: InkWell(
+              borderRadius: BorderRadius.circular((widget.topSelectorStyle.closeIconSize ?? 10) * 5),
+              onTap: () {
+                Navigator.of(context).pop();
+              },
+              child: Padding(
+                padding: const EdgeInsets.all(8.0),
+                child: Icon(
+                  Icons.close_rounded,
+                  color: widget.topSelectorStyle.closeIconColor,
+                  size: widget.topSelectorStyle.closeIconSize,
                 ),
               ),
             ),
-          )
+          ),
         ],
       ),
     );
   }
 
+  // Widget header() {
+  //   return Padding(
+  //     padding: const EdgeInsets.all(10),
+  //     child: Row(
+  //       children: [
+  //         Expanded(
+  //           flex: 1,
+  //           child: Align(
+  //             alignment: FractionalOffset.topCenter,
+  //             child: Padding(
+  //               padding: const EdgeInsets.fromLTRB(0, 5, 20, 0),
+  //               child: Container(),
+  //             ),
+  //           ),
+  //         ),
+  //         Expanded(
+  //           flex: 10,
+  //           child: Align(
+  //             alignment: Alignment.center,
+  //             child: Padding(
+  //               padding: const EdgeInsets.only(top: 5),
+  //               child: Text(
+  //                 widget.hint,
+  //                 textAlign: TextAlign.left,
+  //                 style: TextStyle(
+  //                   color: widget.selectPickerInputSearchStyle.textColor,
+  //                   fontSize: 18,
+  //                   height: 1,
+  //                 ),
+  //               ),
+  //             ),
+  //           ),
+  //         ),
+  //         Expanded(
+  //           flex: 2,
+  //           child: Padding(
+  //             padding: const EdgeInsets.all(8.0),
+  //             child: Align(
+  //               alignment: Alignment.centerRight,
+  //               child: Row(
+  //                 mainAxisAlignment: MainAxisAlignment.center,
+  //                 children: [
+  //                   InkWell(
+  //                     borderRadius: BorderRadius.circular(50),
+  //                     onTap: () {
+  //                       Navigator.pop(context);
+  //                     },
+  //                     child: Center(
+  //                       child: Icon(
+  //                         Icons.close,
+  //                         color: widget.selectPickerBottomSheetStyle.closeIconColor,
+  //                         size: widget.selectPickerBottomSheetStyle.iconSize,
+  //                       ),
+  //                     ),
+  //                   ),
+  //                 ],
+  //               ),
+  //             ),
+  //           ),
+  //         )
+  //       ],
+  //     ),
+  //   );
+  // }
+
   Widget hintSearch() {
     return widget.hintSearch != null
         ? Padding(
-            padding: const EdgeInsets.fromLTRB(30, 0, 30, 15),
+            padding: const EdgeInsets.fromLTRB(30, 15, 30, 0),
             child: SelectPickerInputNoBorder(
               labelColor: widget.selectPickerInputSearchStyle.textColor,
               backgroundColor: widget.selectPickerInputSearchStyle.backgroundColor,